Refactor: Reorganise eventListeners directory
